use std::size_t to avoid compile warning

+++ Standard/ch02/Intvector.cpp
@@ -8,14 +8,15 @@ using namespace std;
 
 int main() {
     vector<int> v;
-    for (int i = 0; i < 10; i++)
+
+    for (size_t i = 0; i < 10; i++)
         v.push_back(i);
-    for (int i = 0; i < v.size(); i++)
+    for (size_t i = 0; i < v.size(); i++)
         cout << v[i] << ", ";
     cout << endl;
-    for (int i = 0; i < v.size(); i++)
+    for (size_t i = 0; i < v.size(); i++)
         v[i] = v[i] * 10; // Assignment
-    for (int i = 0; i < v.size(); i++)
+    for (size_t i = 0; i < v.size(); i++)
         cout << v[i] << ", ";
     cout << endl;
 } ///:~
